+++ app/templates/confirm-choices.html
